requirements.txt
slack-bolt>=1.18.0
psycopg2-binary>=2.9.9
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0

